# Abstract operators to no longer rely on instance vars