--- a/src/pluto/channel_type.cr
+++ b/src/pluto/channel_type.cr
@@ -1,4 +1,4 @@
-enum Pluto::Channel
+enum Pluto::ChannelType
   Red
   Green
   Blue