--- a/src/pluto/operation/bilinear_resize.cr
+++ b/src/pluto/operation/bilinear_resize.cr
@@ -7,7 +7,7 @@ module Pluto::Operation::BilinearResize
     x_ratio = width > 1 ? (@width - 1) / (width - 1) : 0
     y_ratio = height > 1 ? (@height - 1) / (height - 1) : 0
 
-    each_channel do |channel, ch|
+    each_channel do |channel, channel_type|
       resized_channel = Array.new(width * height) { 0u8 }
 
       height.times do |h|
@@ -50,7 +50,7 @@ module Pluto::Operation::BilinearResize
         end
       end
 
-      self[ch] = resized_channel
+      self[channel_type] = resized_channel
     end
 
     @width = width