--- a/src/pluto/operation/brightness.cr
+++ b/src/pluto/operation/brightness.cr
@@ -4,10 +4,10 @@ module Pluto::Operation::Brightness
   end
 
   def brightness!(value : Float64) : Image
-    (@width * @height).times do |index|
-      @red.unsafe_put(index, Math.min(255, (@red.unsafe_fetch(index) * value)).to_u8)
-      @green.unsafe_put(index, Math.min(255, (@green.unsafe_fetch(index) * value)).to_u8)
-      @blue.unsafe_put(index, Math.min(255, (@blue.unsafe_fetch(index) * value)).to_u8)
+    each_channel do |channel|
+      size.times do |index|
+        channel.unsafe_put(index, Math.min(255, (channel.unsafe_fetch(index) * value)).to_u8)
+      end
     end
     self
   end