--- a/src/pluto/operation/channel_swap.cr
+++ b/src/pluto/operation/channel_swap.cr
@@ -1,9 +1,9 @@
 module Pluto::Operation::ChannelSwap
-  def channel_swap(a : Channel, b : Channel) : Image
+  def channel_swap(a : ChannelType, b : ChannelType) : Image
     clone.channel_swap!(a, b)
   end
 
-  def channel_swap!(a : Channel, b : Channel) : Image
+  def channel_swap!(a : ChannelType, b : ChannelType) : Image
     ch_a, ch_b = self[a], self[b]
     self[a] = ch_b
     self[b] = ch_a