I'm interested in information extraction from images at some point (i.e. edge detection, OCR, etc.), and these operations are easier on grey-scale images. I noticed the current set of operations use hard coded references to the `@red`, `@green`, and `@blue` channels whereas a grey-scale image would only have one channel.

This refactor provides some private convenience methods in the `Image` class and removes the hardcoded usage of the channel instance variables. I think without sacrificing performance too (but you might have a better way to measure that than me just rerunning `crystal spec` multiple times :P ).

Sometime in the next week I'm hoping to introduce the `GreyImage` type and conversions between it and the color `Image`, and then I can start adding some more operations from there (if you'll accept them).

Thanks for starting this library!