@sanks64 2023-02-06
Tried in one of my projects:

- Before: 4.34s, 5.39s
- After: 4.32s, 5.37s

Tried with [this script](https://github.com/phenopolis/pluto/pull/10):

|                            | Before - Memory | Before - Time | After - Memory | After - Time |
| -------------------------- | --------------- | ------------- | -------------- | ------------ |
| bilinear_resize!(640, 480) | 1.17MiB         | 6ms           | 1.17MiB        | 6ms          |
| box_blur!(10)              | 600kiB          | 6ms           | 600kiB         | 6ms          |
| brightness!(1.4)           | 0B              | 1ms           | 0B             | 1ms          |
| channel_swap!(:red, :blue) | 0B              | 0ms           | 0B             | 0ms          |
| contrast!(128)             | 0B              | 1ms           | 0B             | 1ms          |
| gaussian_blur!(10)         | 1.75MiB         | 17ms          | 1.75MiB        | 16ms         |
| horizontal_blur!(10)       | 300kiB          | 2ms           | 300kiB         | 2ms          |
| vertical_blur!(10)         | 300kiB          | 3ms           | 300kiB         | 3ms          |

So in terms of performance it's good