@sanks64 2023-02-06
Thanks for the awesome PR :)

Feel free to go ahead with `GreyImage` and new operations as well.