@sanks64 2023-02-06
Btw, forgot to say. Please specify return types where possible.