@Vici37 2023-02-06
If it's alright with you, I can rename `Channel` to `ChannelType` in my next PR? (Sorry, just saw the above comment, and don't want to have you kickoff the github actions again for yet another commit :grimacing: )